updated assignment 3
seperated calculator functionality into another package and added reusability of calculator

+++ assignment3/operations/operations.go
@@ -2,21 +2,21 @@
 package operations
 
 // Add returns the sum of two numbers.
-func Add(numA int, numbB int) int {
+func Add(numA float32, numbB float32) float32 {
 	return numA + numbB
 }
 
 // Sub returns the difference between two numbers.
-func Sub(numA int, numbB int) int {
+func Sub(numA float32, numbB float32) float32 {
 	return numA - numbB
 }
 
 // Mul returns multiplication of two numbers.
-func Mul(numA int, numbB int) int {
+func Mul(numA float32, numbB float32) float32 {
 	return numA * numbB
 }
 
 // Div returns division of two numbers.
-func Div(numA int, numbB int) (int, int) {
-	return numA / numbB, numA % numbB   
+func Div(numA float32, numbB float32) (int, int) {
+	return int(numA) / int(numbB), int(numA) % int(numbB)
 }
